Move private ImageAdapter in grid widgets to own file
